+++ src/main.cpp
@@ -16,14 +16,12 @@
 #include "ros/ros_video_capture.h"
 #include "signal_listener.h"
 #else
-#ifdef __APPLE__
+#if defined(__APPLE__)
 #include "mac_helper/mac_capturer.h"
-#else
-#if USE_MMAL_ENCODER | USE_JETSON_ENCODER
+#elif defined(__linux__)
 #include "v4l2_video_capturer/v4l2_video_capturer.h"
 #else
 #include "rtc/device_video_capturer.h"
-#endif
 #endif
 #endif
 
@@ -80,16 +78,15 @@ int main(int argc, char* argv[]) {
     return 1;
   }
   rtc::LogMessage::AddLogToStream(log_sink.get(), rtc::LS_INFO);
-#ifdef __APPLE__
+  auto size = cs.getSize();
+#if defined(__APPLE__)
   rtc::scoped_refptr<MacCapturer> capturer = MacCapturer::Create(
-      cs.getWidth(), cs.getHeight(), cs.framerate, cs.video_device);
-#else
-#if USE_MMAL_ENCODER || USE_JETSON_ENCODER
+      size.width, size.height, cs.framerate, cs.video_device);
+#elif defined(__linux__)
   rtc::scoped_refptr<V4L2VideoCapture> capturer = V4L2VideoCapture::Create(cs);
 #else
   rtc::scoped_refptr<DeviceVideoCapturer> capturer =
-      DeviceVideoCapturer::Create(cs.getWidth(), cs.getHeight(), cs.framerate);
-#endif
+      DeviceVideoCapturer::Create(size.width, size.height, cs.framerate);
 #endif
   if (!capturer && !cs.no_video) {
     std::cerr << "failed to create capturer" << std::endl;