+++ src/util.cpp
@@ -1,5 +1,8 @@
 #include "util.h"
 
+#include <regex>
+
+// external libraries
 #include <CLI/CLI.hpp>
 #include <boost/beast/version.hpp>
 #include <boost/filesystem/operations.hpp>
@@ -7,15 +10,10 @@
 #include <boost/preprocessor/stringize.hpp>
 #include <nlohmann/json.hpp>
 
+#include "momo_version.h"
 #include "rtc_base/helpers.h"
 #if USE_ROS
 #include "ros/ros.h"
-#endif
-
-// バージョン情報
-// 通常は外から渡すが、渡されていなかった場合の対応
-#ifndef MOMO_VERSION
-#define MOMO_VERSION "internal-build"
 #endif
 
 // HWA を効かせる場合は 1 になる
@@ -91,6 +89,9 @@ void Util::parseArgs(int argc,
                        cs.disable_highpass_filter);
   local_nh.param<bool>("disable_typing_detection", cs.disable_typing_detection,
                        cs.disable_typing_detection);
+  local_nh.param<bool>("disable_residual_echo_detector",
+                       cs.disable_residual_echo_detector,
+                       cs.disable_residual_echo_detector);
 
   if (use_sora && local_nh.hasParam("SIGNALING_URL") &&
       local_nh.hasParam("CHANNEL_ID")) {
@@ -167,6 +168,24 @@ void Util::parseArgs(int argc,
       },
       "");
 
+  auto is_valid_resolution = CLI::Validator(
+      [](std::string input) -> std::string {
+        if (input == "QVGA" || input == "VGA" || input == "HD" ||
+            input == "FHD" || input == "4K") {
+          return std::string();
+        }
+
+        // 数値x数値、というフォーマットになっているか確認する
+        std::regex re("^[1-9][0-9]*x[1-9][0-9]*$");
+        if (std::regex_match(input, re)) {
+          return std::string();
+        }
+
+        return "解像度は QVGA, VGA, HD, FHD, 4K, 幅x高さ "
+               "どれかである必要があります。";
+      },
+      "");
+
   app.add_flag("--no-video", cs.no_video, "ビデオを表示しない");
   app.add_flag("--no-audio", cs.no_audio, "オーディオを出さない");
   app.add_flag("--force-i420", cs.force_i420,
@@ -176,17 +195,18 @@ void Util::parseArgs(int argc,
                "MJPEGのデコードとビデオのリサイズをハードウェアで行う"
                "（対応デバイスのみ）")
       ->check(is_valid_use_native);
-#if USE_MMAL_ENCODER || USE_JETSON_ENCODER
-  app.add_option("--video-device", cs.video_device,
-                 "デバイスファイル名。省略時はどれかのビデオデバイスを自動検出")
-      ->check(CLI::ExistingFile);
-#elif __APPLE__
+#if defined(__APPLE__)
   app.add_option("--video-device", cs.video_device,
                  "デバイス番号、またはデバイス名。省略時はデフォルト（デバイス"
                  "番号が0）のビデオデバイスを自動検出");
+#elif defined(__linux__)
+  app.add_option("--video-device", cs.video_device,
+                 "デバイスファイル名。省略時はどれかのビデオデバイスを自動検出")
+      ->check(CLI::ExistingFile);
 #endif
-  app.add_set("--resolution", cs.resolution, {"QVGA", "VGA", "HD", "FHD", "4K"},
-              "解像度");
+  app.add_option("--resolution", cs.resolution,
+                 "解像度(QVGA, VGA, HD, FHD, 4K, 幅x高さ)")
+      ->check(is_valid_resolution);
   app.add_option("--framerate", cs.framerate, "フレームレート")
       ->check(CLI::Range(1, 60));
   app.add_flag("--fixed-resolution", cs.fixed_resolution, "固定解像度");
@@ -222,6 +242,9 @@ void Util::parseArgs(int argc,
                "ハイパスフィルター無効");
   app.add_flag("--disable-typing-detection", cs.disable_typing_detection,
                "タイピングディテクション無効");
+  app.add_flag("--disable-residual-echo-detector",
+               cs.disable_residual_echo_detector,
+               "残響エコーディテクション無効");
 
   auto test_app = app.add_subcommand("test", "開発向け");
   auto ayame_app = app.add_subcommand("ayame", "WebRTC Signaling Server Ayame");
@@ -300,8 +323,8 @@ void Util::parseArgs(int argc,
   }
 
   if (version) {
-    std::cout << "WebRTC Native Client Momo version " MOMO_VERSION
-                 " USE_MMAL_ENCODER=" BOOST_PP_STRINGIZE(MOMO_USE_MMAL_ENCODER)
+    std::cout << MOMO_NAME
+        " USE_MMAL_ENCODER=" BOOST_PP_STRINGIZE(MOMO_USE_MMAL_ENCODER)
               << std::endl;
     exit(0);
   }